requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djangorestframework_simplejwt==5.1.0
requests==2.22.0
sshtunnel==0.4.0
django-cors-headers==3.13.0
mysqlclient==2.1.0
django-cors-headers==3.13.0
pymysql==1.1.0
openpyxl==3.0.10
pillow==9.5.0
python-decouple==3.6
